refactor: move Constants to root package
