+++ imx-core-sdk-android/src/main/java/com/immutable/sdk/model/EthAsset.kt
@@ -3,7 +3,7 @@ package com.immutable.sdk.model
 import com.immutable.sdk.api.model.SignableToken
 import com.immutable.sdk.api.model.Token
 import com.immutable.sdk.api.model.TokenData
-import com.immutable.sdk.utils.Constants
+import com.immutable.sdk.Constants
 
 /**
  * This is an ETH wrapper for the [Token] API model
